Finish day 14
I think there is a bug because the answer is different if counts_after_n
is used after step_n rather than just using counts_after_n. Step,
step_n, and counts are all vestiges of a previous attempt that are
valid, but far to slow to run with much more than 10 steps.